eventbuilding: Merge config reading into one class
Before, the reading of the configuration options was done
through sperate functions and with global variables.
This is now merged into a single EcalConfig class.

This commit should only have one functionality change:
It is now possible to specify the parent directory of all config files.
Some apparent bugs and improvement opportunities ere identified.
These are NOT change in this commit, in order to not hide
their change under the large-scale refactoring that is performed here.