Do commissioning once at start instead of per-hit
This trades a rather small increase in startup cost in
for a 5-times speedup in the per-event hit building.
Overall, this translates into ~20% performance increase
on a small dataset (30000 hits).

This is a prerequisit for getting rid of the loop over each EcalHit.
That change will be introduced in an upcoming commit.