Special-case negative fit errors
Special-case negative fit errors are (currently) indicating
that the fit (for the pedestal/mip) was problematic.
The value of the negative error then indicates the type of problem.
The values in the calibration file are then replaced by an average.

It is preferred to do this averaging (independently) in the
eventbuilding procedure.
This workflow already exists, and does toggle the isCommissioned
variable properly.
By setting the values with negative-error to zero, the
eventbuilding procedure knows that it has to replace the value.