Eventbuilding: Allow merging within chip
So far, merging of different BCIDs only merged separate chips.
It was found that within a shower, often the hits are spread out
over more than one BCID within the same chip. I am not sure it is
_supposed to be like this_, but it seems to be happening quite
frequently.
On a prelimarily calibrated 3GeV run, the energy spread was
reduced from 35% to 22% (before any processing apart from
requiring min_chips_with_hits >= 20). I believe this shows that
the change is both necessary and going in the right direction.

Still, merging within the same chip seems shady. Thus we introduce
a steering option to toggle this on/off.