Update event branches
hg/lg are replaced with adc_high/low for consistency with
converted.root. Where lg is used to specif the "mode" e.g. energy
vs energy_lg, it is kept as is.
sum_hg is removed. With different wafer and tungsten thickness per
layer, this variable is meaningless. As I have seen this being
misinterpreted, it is better to remove it.
bcid_prev/bcid_next have little use: just check the bcid branch of
event+/-1 within the same cycle.